+++ src/eleventy/filters/get-style-path.js
@@ -1,0 +1,5 @@
+module.exports = function (eleventyConfig) {
+  eleventyConfig.addFilter("getStylesPath", function (inputPath) {
+    return inputPath.split("/").slice(1, -1).join("/") + "/style.css";
+  });
+}
